refactor merge actions
Merge action starts to become big.

This change split the file in 3 parts: action, queue, helpers

action.py contains the action itself
queue.py all queue manipulation and the celery task
helpers.py common needs of the two others
