+++ setup.cfg
@@ -65,7 +65,7 @@ mergify_actions =
     backport = mergify_engine.actions.backport:BackportAction
     label = mergify_engine.actions.label:LabelAction
     delete_head_branch = mergify_engine.actions.delete_head_branch:DeleteHeadBranchAction
-    merge = mergify_engine.actions.merge:MergeAction
+    merge = mergify_engine.actions.merge.action:MergeAction
     close = mergify_engine.actions.close:CloseAction
     comment = mergify_engine.actions.comment:CommentAction
     dismiss_reviews = mergify_engine.actions.dismiss_reviews:DismissReviewsAction
